+++ c10/core/CPUAllocator.h
@@ -1,40 +1,20 @@
 #pragma once
 
-#include <cstring>
 #include <unordered_map>
 
 #include <c10/core/Allocator.h>
-#include <c10/util/Logging.h>
-#include <c10/util/numa.h>
+#include <c10/macros/Macros.h>
+#include <c10/util/Flags.h>
 
 // TODO: rename to c10
 C10_DECLARE_bool(caffe2_report_cpu_memory_usage);
-C10_DECLARE_bool(caffe2_cpu_allocator_do_zero_fill);
-C10_DECLARE_bool(caffe2_cpu_allocator_do_junk_fill);
 
 namespace c10 {
-
-#ifdef C10_MOBILE
-// Use 16-byte alignment on mobile
-// - ARM NEON AArch32 and AArch64
-// - x86[-64] < AVX
-constexpr size_t gAlignment = 16;
-#else
-// Use 64-byte alignment should be enough for computation up to AVX512.
-constexpr size_t gAlignment = 64;
-#endif
 
 using MemoryDeleter = void (*)(void*);
 
 // A helper function that is basically doing nothing.
 C10_API void NoDelete(void*);
-
-// Fill the data memory region of num bytes with a particular garbage pattern.
-// The garbage value is chosen to be NaN if interpreted as floating point value,
-// or a very large integer.
-C10_API void memset_junk(void* data, size_t num);
-
-C10_API void* alloc_cpu(size_t nbytes);
 
 // A simple struct that is used to report C10's memory allocation and
 // deallocation status to the profiler
