+++ c10/mobile/CPUProfilingAllocator.cpp
@@ -1,5 +1,6 @@
 #include <climits>
 
+#include <c10/core/alloc_cpu.h>
 #include <c10/core/free_cpu.h>
 #include <c10/mobile/CPUProfilingAllocator.h>
 #include <c10/util/irange.h>
